% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{binfactor}
\alias{binfactor}
\title{binfactor}
\usage{
binfactor(x, ...)
}
\arguments{
\item{vector}{of data to be binned}
}
\value{
vector of binned data
}
\description{
binfactor
}
\examples{
water_type <- floor(runif(100,1,10))
water_type <- dhsdata:::binfactor(water_type, 1:5, 6:8, 9:10)
# the first vetor corresponds to bin = 0, the second vector corresponds to bin = 1 and so on...

}
