% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths_default.R
\name{default_data_set_path}
\alias{default_data_set_path}
\title{Get the default data set path}
\usage{
default_data_set_path()
}
\value{
installed package default data paths
}
\description{
Get the default data set path
}
